created a new class to init FileSystemHelpers instead of main
